pyproject.toml
[project]
name = "echoframe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.1.4",
    "django-import-export>=4.3.3",
    "pillow>=11.0.0",
    "python-dotenv>=1.0.1",
    "tzdata>=2024.2",
]

[dependency-groups]
dev = [
    "djlint>=1.36.4",
    "ruff>=0.8.4",
]
